moved metrics into a new namespace
