+++ CardPeak.Core/Repository/IApprovalTransactionAgentRepository.cs
@@ -1,4 +1,5 @@
 ﻿using CardPeak.Domain;
+using CardPeak.Domain.Metrics;
 using System;
 using System.Collections.Generic;
