+++ CardPeak.Core/Service/IDashboardService.cs
@@ -1,4 +1,5 @@
 ﻿using CardPeak.Domain;
+using CardPeak.Domain.Metrics;
 using System.Collections.Generic;
 
 namespace CardPeak.Core.Service
