+++ CardPeak.Domain/Metrics/AgentRankMetric.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
-namespace CardPeak.Domain
+namespace CardPeak.Domain.Metrics
 {
     public sealed class AgentRankMetric : ApprovalMetric<Agent>
     {
