+++ CardPeak.Repository.EF/ApprovalTransactionAgentRepository.cs
@@ -1,6 +1,7 @@
 ﻿using CardPeak.Core.Repository;
 using CardPeak.Domain;
 using CardPeak.Domain.Constants;
+using CardPeak.Domain.Metrics;
 using System;
 using System.Collections.Generic;
 using System.Data.Entity;
