+++ CardPeak.Service/MetricsService.cs
@@ -1,6 +1,6 @@
 ﻿using CardPeak.Core.Repository;
 using CardPeak.Core.Service;
-using CardPeak.Domain;
+using CardPeak.Domain.Metrics;
 using CardPeak.Repository.EF;
 using System;
 using System.Collections.Generic;
